refactor: export API package
