+++ api/skillpath.go
